Fix map download resumption using exponential backoff
The Android MapboxSDK OfflineManager does not directly enable the com.mapbox.mapboxsdk.net.ConnectivityReceiver when offline region download resumes. It is only enabled when an offline region is being created and deactivated immediately after. However, the OfflineManager and it's components do use the com.mapbox.mapboxsdk.net.ConnectivityReceiver(which is a singleton) if it is activated.

This fix activates the com.mapbox.mapboxsdk.net.ConnectivityReceiver before an OfflineRegion status is activated and deactivates after the OfflineRegion is done downloading, encounters a tile count limit exception or the Service is shutdown