Remove re-adding layers when style changes
There is no default style for the MapView and therefore the host application requires to define it explicitly. This means that KujakuMapView will always receive the callback when the host application does this. This is because the host application needs to wait for the MapView to load before setting the style